Move `IsValid` under test fixtures
